refactor: :recycle: Refactor after CR
