+++ runbotics/runbotics-orchestrator-ui/src/src-app/components/Tile/ProcessTile/ProcessTile.tsx
@@ -166,12 +166,12 @@ const ProcessTile: FC<ProcessTileProps> = ({ process }) => {
             return;
         }
 
-        const job = jobsMap[processId];
-        if (!job) return;
+        const jobPayload = { ...jobsMap[processId] };
+        if (!jobPayload) return;
 
-        const eventType = job?.eventType;
+        const eventType = jobPayload?.eventType;
         if (eventType === WsMessage.JOB_WAITING) {
-            dispatch(schedulerActions.removeWaitingJob({ jobId: String(job.jobId) }))
+            dispatch(schedulerActions.removeWaitingJob({ jobId: String(jobPayload.jobId) }))
                 .then(() => {
                     enqueueSnackbar(translate(
                         'Component.BotProcessRunner.Success.RemovedJob',
@@ -218,11 +218,11 @@ const ProcessTile: FC<ProcessTileProps> = ({ process }) => {
     }, [processInstance]);
 
     useEffect(() => {
-        const jobPayload = jobsMap[processId];
+        const jobPayload = { ...jobsMap[processId] };
         if (!jobPayload) return;
 
         switch (jobPayload.eventType) {
-            case WsMessage.PROCESS_START_COMPLETED:
+            case WsMessage.PROCESS_STARTED:
                 handleRunCompleted();
                 break;
             case WsMessage.JOB_FAILED:
