+++ runbotics/runbotics-orchestrator-ui/src/src-app/translations/pl/process/process.json
@@ -17,7 +17,7 @@
     "Process.Instance.Status.Terminated": "PRZERWANY",
     "Process.Instance.Status.Completed": "UKOŃCZONY",
     "Process.Instance.Status.Queued": "W KOLEJCE",
-    "Process.Instance.Status.Pending": "PRZETWARZANIE",
+    "Process.Instance.Status.Pending": "OCZEKUJĄCY",
     "Process.ProcessInstance.Details.Header.Dialog.Title": "Wynik procesu",
     "Process.ProcessInstance.Details.Header.Dialog.InvalidOutputType": "Wybrany typ wyniku procesu jest niekompatybilny z typem danych do wyświetlenia.",
     "Process.ProcessInstance.Details.Header.Button": "Pokaż wynik",
